requirements.txt
deepgram-sdk==5.3.2
fastapi==0.115.0
python-dotenv==1.0.1
toml==0.10.2
uvicorn[standard]==0.32.0
